Add GenericTool class and update GenericToolsAgent
